refactor(tests): add test_challenges_get_chips_challenge.py
